about created_by
created_by equals models which is a foreighn key and the foreignkey object or method has a User, on_delete=models.CASCADE, and related_name='product_creator' as parameters. This has all the users that i might set up in this django app. A foreign is built to that table that django builds automatically and the person who makes that product is recorded. The related name is 'product_creator'