The Meta class with 'ordering=('-created') or created with out the -
the ordering variable we can order when we return all the data from the Product class or Product table, we can specify what order we want that data to be in. The 'created' without the - is ascending order and with the - is descending order. We specify how it is ordered with the created field or created variable. The last item added to the database gets shown first