Add link from urls/core to urls/store in urls/core, add code to views.py
In views.py Category and Product models or tables are added to the views.py file then the all_products(request) function is created which adds all the Products from model to the products variable then at the bottom the function returns a rendered function which has the request, store/home.html parameters. Rendering is the process of gathering the data and load the associated template, so get the data and the template and send it back to the user. The next step is to make a new 'template' folder and inside of that folder create a store folder and inside of store folder create a home.html file.